Add files via upload
The code performs the processing and analysis of the annual PV installation dataset, starting with the loading and normalisation of the weather and power output variables from the Excel file containing the recorded data from the actual installation and the weather station. The data is then converted into tensors, divided into training and validation sets, and a simple multi-layer neural network is defined with PyTorch, which is trained to predict the power generated from the weather conditions. During training, the early stopping technique is applied to avoid overfitting by storing the best model found. Finally, the performance of the network is evaluated using metrics such as MSE, RMSE, MAE and R², comparing the results obtained by the neural network with those of a reference mathematical model. In addition, the scaling transformer is saved for future use and the results of the model comparison are printed. For the MLP, LSTM and GRU model.